Super Commit

+++ idorm/Source/Presentation/Community/Reactors/PostingViewReactor.swift
@@ -10,9 +10,15 @@ import UIKit
 import ReactorKit
 import RxMoya
 import Photos
+import Kingfisher
 
 final class PostingViewReactor: Reactor {
-    
+  
+  struct Photo: Equatable {
+    let image: UIImage
+    let index: Int
+  }
+  
   enum Action {
     case didTapPictIv
     case didTapCompleteBtn
@@ -21,11 +27,12 @@ final class PostingViewReactor: Reactor {
     case didChangeTitle(String)
     case didChangeContent(String)
     case didTapAnonymousBtn(Bool)
+    case viewDidLoad
   }
   
   enum Mutation {
     case setGalleryVC(Bool)
-    case setImages([UIImage])
+    case setImages([Photo])
     case deleteImages(Int)
     case setTitle(String)
     case setContents(String)
@@ -33,11 +40,13 @@ final class PostingViewReactor: Reactor {
     case setAnonymous(Bool)
     case setLoading(Bool)
     case setPopVC(Bool)
+    case setDeletePostPhotoIds(Int)
   }
   
   struct State {
     var showsGalleryVC: Bool = false
-    var currentImages: [UIImage] = []
+    var currentImages: [Photo] = []
+    var deletePostPhotoIds: [Int] = []
     var currentTitle: String = ""
     var currentContents: String = ""
     var isEnabledCompleteBtn: Bool = false
@@ -54,13 +63,14 @@ final class PostingViewReactor: Reactor {
   private let currentDorm: Dormitory
   var post: CommunityResponseModel.Post?
   var initialState: State = State()
+  let postingType: PostingType
   
   init(
     _ postingType: PostingType,
     dorm: Dormitory
   ) {
     self.currentDorm = dorm
-    
+    self.postingType = postingType
     switch postingType {
     case .new:
       break
@@ -71,26 +81,63 @@ final class PostingViewReactor: Reactor {
   
   func mutate(action: Action) -> Observable<Mutation> {
     switch action {
+    case .viewDidLoad:
+      switch postingType {
+      case .new:
+        return .empty()
+      case .edit:
+        guard let post = post else { return .empty() }
+        var photos: [Photo] = []
+        
+        post.postPhotos.forEach { [weak self] in
+          var image: UIImage?
+          self?.downloadImage(
+            from: URL(string: $0.photoUrl)!
+          ) {
+            image = $0
+          }
+          photos.append(Photo(image: image!, index: $0.photoId))
+        }
+        
+        return .concat([
+          .just(.setTitle(post.title)),
+          .just(.setContents(post.content)),
+          .just(.setImages(photos)),
+          .just(.setAnonymous(post.isAnonymous))
+        ])
+      }
+      
     case .didTapPictIv:
       return .concat([
         .just(.setGalleryVC(true)),
         .just(.setGalleryVC(false))
       ])
+      
     case .didTapCompleteBtn:
-      let newPost = CommunityRequestModel.Post(
-        content: currentState.currentContents,
-        title: currentState.currentTitle,
-        dormCategory: currentDorm,
-        images: currentState.currentImages,
-        isAnonymous: currentState.isAnonymous
-      )
-      return .concat([
-        .just(.setLoading(true)),
-        CommunityAPI.provider.rx.request(.savePost(newPost))
+      switch postingType {
+      case .edit:
+        guard let post = post else { return .empty() }
+        let photos = currentState.currentImages
+          .filter { $0.index < 0 }
+          .map { $0.image }
+        
+        let newPost = CommunityRequestModel.Post(
+          content: currentState.currentContents,
+          title: currentState.currentTitle,
+          dormCategory: .no1,
+          images: photos,
+          isAnonymous: currentState.isAnonymous
+        )
+        
+        return .concat([
+          .just(.setLoading(true)),
+          CommunityAPI.provider.rx.request(.editPost(
+            postId: post.postId,
+            post: newPost,
+            deletePostPhotos: currentState.deletePostPhotoIds)
+          )
           .asObservable()
-          .retry()
-          .withUnretained(self)
-          .flatMap { owner, response -> Observable<Mutation> in
+          .flatMap { response -> Observable<Mutation> in
             switch response.statusCode {
             case 200..<300:
               return .concat([
@@ -98,14 +145,65 @@ final class PostingViewReactor: Reactor {
                 .just(.setPopVC(true))
               ])
             default:
-              fatalError("게시글 저장 실패")
+              fatalError()
             }
           }
-      ])
+        ])
+        
+      case .new:
+        var images: [UIImage] = []
+        currentState.currentImages.forEach {
+          images.append($0.image)
+        }
+        let newPost = CommunityRequestModel.Post(
+          content: currentState.currentContents,
+          title: currentState.currentTitle,
+          dormCategory: currentDorm,
+          images: images,
+          isAnonymous: currentState.isAnonymous
+        )
+        return .concat([
+          .just(.setLoading(true)),
+          CommunityAPI.provider.rx.request(.savePost(newPost))
+            .asObservable()
+            .retry()
+            .withUnretained(self)
+            .flatMap { owner, response -> Observable<Mutation> in
+              switch response.statusCode {
+              case 200..<300:
+                return .concat([
+                  .just(.setLoading(false)),
+                  .just(.setPopVC(true))
+                ])
+              default:
+                fatalError("게시글 저장 실패")
+              }
+            }
+        ])
+      }
     case .didPickedImages(let images):
-      return .just(.setImages(images))
+      var photos: [Photo] = []
+      images.forEach {
+        photos.append(Photo(image: $0, index: -1))
+      }
+      return .just(.setImages(photos))
+      
     case .didTapDeleteBtn(let index):
-      return .just(.deleteImages(index))
+      switch postingType {
+      case .new:
+        return .just(.deleteImages(index))
+      case .edit:
+        guard let post = post else { return .empty() }
+        let photo = currentState.currentImages[index]
+        let deleteIndex = post.postPhotos.first(where: { $0.photoId == photo.index })?.photoId
+        if deleteIndex != nil {
+          return .concat([
+            .just(.deleteImages(index)),
+            .just(.setDeletePostPhotoIds(deleteIndex!))
+          ])
+        }
+        return .just(.deleteImages(index))
+      }
     case .didChangeTitle(let title):
       return .concat([
         .just(.setTitle(title)),
@@ -159,8 +257,26 @@ final class PostingViewReactor: Reactor {
       
     case .setPopVC(let state):
       newState.popVC = state
+      
+    case .setDeletePostPhotoIds(let id):
+      newState.deletePostPhotoIds.append(id)
     }
     
     return newState
+  }
+}
+
+extension PostingViewReactor {
+  // 이미지를 다운로드하고 이를 UIImage로 변환하는 함수
+  func downloadImage(from url: URL, completion: @escaping (UIImage?) -> Void) {
+    KingfisherManager.shared.retrieveImage(with: url) { result in
+      switch result {
+      case .success(let imageResult):
+        completion(imageResult.image)
+      case .failure(let error):
+        print("Error: \(error.localizedDescription)")
+        completion(nil)
+      }
+    }
   }
 }
